# src/main/java/com/example/demo/controller/AddressController.java
package com.example.demo.controller;

import static org.springframework.http.ResponseEntity.notFound;

import java.util.List;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import com.example.demo.AddressApi;
import com.example.demo.hateoas.AddressRepresentationModelAssembler;
import com.example.demo.model.AddAddressReq;
import com.example.demo.model.Address;
import com.example.demo.service.AddressService;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RequiredArgsConstructor
@RestController
public class AddressController implements AddressApi {

  private final AddressService service;
  private final AddressRepresentationModelAssembler assembler;

  @Override
  public ResponseEntity<Address> createAddress(@Valid AddAddressReq addAddressReq) {
    return ResponseEntity
        .status(HttpStatus.CREATED)
        .body(service.createAddress(addAddressReq).map(assembler::toModel).get());
  }

  @Override
  public ResponseEntity<Void> deleteAddressesById(String id) {
    service.deleteAddressesById(id);
    return ResponseEntity.accepted().build();
  }

  @Override
  public ResponseEntity<Address> getAddressesById(String id) {
    return service.getAddressesById(id)
        .map(assembler::toModel)
        .map(ResponseEntity::ok)
        .orElse(notFound().build());
  }

  @Override
  public ResponseEntity<List<Address>> getAllAddresses() {
    return ResponseEntity.ok(assembler.toListModel(service.getAllAddresses()));
  }
}